+++ NEWS.md
@@ -1,3 +1,9 @@
+# PMXFrem (development version)
+
+* Made it possible to use availCov to specify the covariates to be used for the derivation of TOTCOVVAR.
+* Changed the default in plotExplainedVar so tha the main effects ordering is based on the mean instead of the median.
+* Added the argument reordFun to specify the reordering function for the main effects ordering.
+    
 # PMXFrem 1.2.5
 
 * Added traceplot() and fixed a bug in createFFEMdata().